Move shared logic to library file
